requirements.txt
Django==4.2.2
django-cors-headers==4.1.0
django-environ==0.10.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
psycopg2==2.9.6
axios==0.4.0